1.16 update, cache rework, and hash with uuids

+++ src/main/java/tut/simpleshoppingdistrict/utils/JSONUtils.java
@@ -1,19 +1,27 @@
 package tut.simpleshoppingdistrict.utils;
 
 
+import com.google.common.collect.Sets;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.reflect.TypeToken;
 import tut.simpleshoppingdistrict.data.SSDRegion;
 
+import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.nio.file.Files;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Logger;
@@ -34,11 +42,19 @@ public class JSONUtils {
 
     //Translating object data to JSON
     private static ArrayList<SSDRegion> jsonStringToSSDRegionArrayList(String jsonString) {
-        Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
-        Type SSDRegionListType = new TypeToken<ArrayList<SSDRegion>>() {
+        final Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
+        final Type SSDRegionListType = new TypeToken<ArrayList<SSDRegion>>() {
         }.getType();
 
         return gson.fromJson(jsonString, SSDRegionListType);
+    }
+
+    //Translating object data to JSON
+    private static HashMap<Long, HashSet<String>> jsonStringToChunkCache(String jsonString) {
+        final Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
+        final Type completeChunkCache = new TypeToken<HashMap<Long, HashSet<String>>>() {}.getType();
+
+        return gson.fromJson(jsonString, completeChunkCache);
     }
 
     //Read Json file from the cache
@@ -61,32 +77,90 @@ public class JSONUtils {
     }
 
     //Save Json file from the cache
-    public static void saveCacheData(ConcurrentHashMap<String, TreeSet<SSDRegion>> cacheData) {
+    public static void savePlayerCacheData(ConcurrentHashMap<String, TreeSet<SSDRegion>> cacheData) {
+        HashMap<Long, HashSet<String>> chunkPlayerClaimContainer = new HashMap<>();
+
         //For each entry of cacheData it serializes the list to Json
         for (Map.Entry<String, TreeSet<SSDRegion>> cacheDataEntry : cacheData.entrySet()) {
-            logger.info("Saving cache data for player " + cacheDataEntry.getKey() + ".");
             String currentUUID = cacheDataEntry.getKey();
             TreeSet<SSDRegion> currentRegionList = cacheDataEntry.getValue();
+            List<Long> chunkHashesInCurrentRegionList = new ArrayList<>();
 
-            File filename = new File(SSDConstants.BASE_PLUGIN_FOLDER_PATH + File.separator + currentUUID + ".json");
+            for (SSDRegion region : currentRegionList) {
+                chunkHashesInCurrentRegionList.addAll(region.getChunkContainerHash());
+            }
+
+            chunkHashesInCurrentRegionList.forEach( entry -> {
+                                              if (chunkPlayerClaimContainer.containsKey(entry)) {
+                                                  HashSet<String> currentSet = chunkPlayerClaimContainer.get(entry);
+                                                  currentSet.add(currentUUID);
+                                                  chunkPlayerClaimContainer.put(entry, currentSet);
+                                              }
+                                              else {
+                                                  chunkPlayerClaimContainer.put(entry, new HashSet<>(Arrays.asList(currentUUID)));
+                                              }
+                                          });
+
+            File filename = new File(SSDConstants.BASE_PLUGIN_FOLDER_PATH + File.separator + "playerdata" + File.separator + currentUUID + ".json");
 
             try {
-
                 if (!filename.exists()) {
                     if (!(filename.getParentFile().mkdirs() && filename.createNewFile())) {
                         logger.info("Created region file for " + currentUUID + ".");
                     }
                 }
 
-                BufferedWriter writer = new BufferedWriter(new FileWriter(filename));
-                writer.write(objectToJsonString(new ArrayList<>(currentRegionList)));
-                writer.flush();
-                writer.close();
+                TreeSet<SSDRegion> fileRegions = new TreeSet<>(JSONUtils.jsonStringToSSDRegionArrayList(new String(Files.readAllBytes(filename.toPath()))));
+
+                if (!fileRegions.equals(currentRegionList)) {
+                    logger.info("Saving cache data for player " + cacheDataEntry.getKey() + ".");
+                    BufferedWriter writer = new BufferedWriter(new FileWriter(filename));
+                    writer.write(objectToJsonString(new ArrayList<>(currentRegionList)));
+                    writer.flush();
+                    writer.close();
+                }
+
 
             } catch (IOException e) {
                 logger.warning("Exception while saving player data!");
-                logger.warning(e.getMessage());
+                e.printStackTrace();
             }
+        }
+        try {
+            File filename = new File(SSDConstants.BASE_PLUGIN_FOLDER_PATH + File.separator + "chunkdata.json");
+            HashMap<Long, HashSet<String>> fileRegions;
+
+            if (!filename.exists()) {
+                if (!(filename.getParentFile().mkdirs() && filename.createNewFile())) {
+                    logger.info("Created new chunkdata.json");
+                }
+
+                BufferedWriter writer = new BufferedWriter(new FileWriter(filename));
+                fileRegions = new HashMap<>();
+                writer.write(objectToJsonString(fileRegions));
+                writer.flush();
+            }
+            else {
+                HashMap<Long, HashSet<String>> contents = JSONUtils.jsonStringToChunkCache(new String(Files.readAllBytes(filename.toPath())));
+                if (contents != null) {
+                    fileRegions = contents;
+                }
+                else {
+                    fileRegions = new HashMap<>();
+                }
+            }
+
+            if (!fileRegions.equals(chunkPlayerClaimContainer)) {
+                BufferedWriter writer = new BufferedWriter(new FileWriter(filename));
+                logger.info("Saving chunk cache data");
+                writer.write(objectToJsonString(chunkPlayerClaimContainer));
+                writer.flush();
+                writer.close();
+            }
+
+        } catch (IOException e) {
+            logger.warning("Exception while saving player data!");
+            e.printStackTrace();
         }
     }
